Add Options For Marking "Today"
Add settable color and flags for marking today (as in the actual current date) as opposed to the currently selected date.

Default is to mark the selected date and not today (to match previous behavior), but this can be changed to mark neither, both, one, or the other. In the case where both are marked and coincide with the same day, the "selected" color will be displayed.